+++ source/components/Checkbox.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import SkinnableComponent from './SkinnableComponent';
+
+export default class Checkbox extends SkinnableComponent {
+  static propTypes = Object.assign({}, SkinnableComponent.propTypes, {});
+}