+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+## vNext
+
+### Features
+
+- Add checkbox component
+
 ## 0.2.7
 
 ### Fixes